+++ packages/crypto/src/sha.ts
@@ -26,6 +26,11 @@ export class Sha1 implements HashFunction {
   }
 }
 
+/** Convenience function equivalent to `new Sha1(data).digest()` */
+export function sha1(data: Uint8Array): Uint8Array {
+  return new Sha1(data).digest();
+}
+
 export class Sha256 implements HashFunction {
   public readonly blockSize = 512 / 8;
 
@@ -49,6 +54,11 @@ export class Sha256 implements HashFunction {
   }
 }
 
+/** Convenience function equivalent to `new Sha256(data).digest()` */
+export function sha256(data: Uint8Array): Uint8Array {
+  return new Sha256(data).digest();
+}
+
 export class Sha512 implements HashFunction {
   public readonly blockSize = 1024 / 8;
 
@@ -70,4 +80,9 @@ export class Sha512 implements HashFunction {
   public digest(): Uint8Array {
     return new Uint8Array(this.impl.digest());
   }
+}
+
+/** Convenience function equivalent to `new Sha512(data).digest()` */
+export function sha512(data: Uint8Array): Uint8Array {
+  return new Sha512(data).digest();
 }