+++ packages/crypto/types/keccak.d.ts
@@ -6,3 +6,5 @@ export declare class Keccak256 implements HashFunction {
   update(data: Uint8Array): Keccak256;
   digest(): Uint8Array;
 }
+/** Convenience function equivalent to `new Keccak256(data).digest()` */
+export declare function keccak256(data: Uint8Array): Uint8Array;